feat(#72): Turn 타이머 기능 구현
