+++ src/components/game/TimerBox.jsx
@@ -7,6 +7,7 @@ import { useTurnStore } from '@/store/turn';
 
 const TimerBox = ({ index }) => {
   const { nextStep } = useStepStore();
+  const { turn, playerCount, resetTurn } = useTurnStore();
 
   const currentTime = index > 0 && index < 7 && TIMER_OPTION[index];
   const initialTime = currentTime.time;
@@ -26,7 +27,16 @@ const TimerBox = ({ index }) => {
     if (index === 0 || index === 7) {
       stopTimer(0);
     }
-  }, [initialTime, index]);
+
+    if (index === 2 && turn === playerCount) {
+      stopTimer(0);
+      nextStep();
+      resetTurn();
+    }
+
+    resetTimer(initialTime);
+    startTimer();
+  }, [initialTime, index, turn]);
 
   return (
     <TimerContainer index={index}>
